+++ src/display/inline.rs
@@ -3,7 +3,7 @@
 use crate::{
     display::context::{calculate_after_context, calculate_before_context, opposite_positions},
     display::style::{self, apply_colors},
-    hunks::Hunk,
+    display::hunks::Hunk,
     lines::{format_line_num, MaxLine},
     options::DisplayOptions,
     parse::syntax::MatchedPos,
