+++ src/display/mod.rs
@@ -1,4 +1,5 @@
 pub mod context;
+pub mod hunks;
 pub mod inline;
 pub mod side_by_side;
 pub mod style;
